+++ community-care-eligibility-api/src/main/java/gov/va/api/health/communitycareeligibility/swaggerexamples/SwaggerErrorResponse.java
@@ -1,4 +1,4 @@
-package gov.va.api.health.communitycareeligibility.api.swaggerexamples;
+package gov.va.api.health.communitycareeligibility.swaggerexamples;
 
 import gov.va.api.health.communitycareeligibility.api.ErrorResponse;
