+++ community-care-eligibility-api/src/test/java/gov/va/api/health/communitycareeligibility/api/CommunityCareEligibilityResponseTest.java
@@ -2,7 +2,7 @@ package gov.va.api.health.communitycareeligibility.api;
 
 import static gov.va.api.health.communitycareeligibility.api.RoundTrip.assertRoundTrip;
 
-import gov.va.api.health.communitycareeligibility.api.swaggerexamples.SwaggerCommunityCareEligibilityResponse;
+import gov.va.api.health.communitycareeligibility.swaggerexamples.SwaggerCommunityCareEligibilityResponse;
 import org.junit.jupiter.api.Test;
 
 public class CommunityCareEligibilityResponseTest {
